Split out test location
