--- ColumnSelector/AboutBox.cs
@@ -1,134 +1,0 @@
-﻿// -----------------------------------------------------------------------
-// <copyright file="AboutBox.cs" company="Ace Olszowka">
-// Copyright (c) Ace Olszowka 2014. All rights reserved.
-// </copyright>
-// -----------------------------------------------------------------------
-
-namespace ColumnSelector
-{
-    using System.Reflection;
-    using System.Windows.Forms;
-
-    /// <summary>
-    /// The About Box.
-    /// </summary>
-    public partial class AboutBox : Form
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="AboutBox"/> class.
-        /// </summary>
-        public AboutBox()
-        {
-            this.InitializeComponent();
-            this.Text = string.Format("About {0}", this.AssemblyTitle);
-            this.labelProductName.Text = this.AssemblyProduct;
-            this.labelVersion.Text = string.Format("Version {0}", this.AssemblyVersion);
-            this.labelCopyright.Text = this.AssemblyCopyright;
-            this.labelCompanyName.Text = this.AssemblyCompany;
-            this.textBoxDescription.Text = this.AssemblyDescription;
-        }
-
-        #region Assembly Attribute Accessors
-
-        /// <summary>
-        /// Gets the AssemblyTitle.
-        /// </summary>
-        public string AssemblyTitle
-        {
-            get
-            {
-                object[] attributes = Assembly.GetExecutingAssembly().GetCustomAttributes(typeof(AssemblyTitleAttribute), false);
-                if (attributes.Length > 0)
-                {
-                    AssemblyTitleAttribute titleAttribute = (AssemblyTitleAttribute)attributes[0];
-                    if (titleAttribute.Title != string.Empty)
-                    {
-                        return titleAttribute.Title;
-                    }
-                }
-
-                return System.IO.Path.GetFileNameWithoutExtension(Assembly.GetExecutingAssembly().CodeBase);
-            }
-        }
-
-        /// <summary>
-        /// Gets the AssemblyVersion.
-        /// </summary>
-        public string AssemblyVersion
-        {
-            get
-            {
-                return Assembly.GetExecutingAssembly().GetName().Version.ToString();
-            }
-        }
-
-        /// <summary>
-        /// Gets the AssemblyDescription.
-        /// </summary>
-        public string AssemblyDescription
-        {
-            get
-            {
-                object[] attributes = Assembly.GetExecutingAssembly().GetCustomAttributes(typeof(AssemblyDescriptionAttribute), false);
-                if (attributes.Length == 0)
-                {
-                    return string.Empty;
-                }
-
-                return ((AssemblyDescriptionAttribute)attributes[0]).Description;
-            }
-        }
-
-        /// <summary>
-        /// Gets the AssemblyProduct.
-        /// </summary>
-        public string AssemblyProduct
-        {
-            get
-            {
-                object[] attributes = Assembly.GetExecutingAssembly().GetCustomAttributes(typeof(AssemblyProductAttribute), false);
-                if (attributes.Length == 0)
-                {
-                    return string.Empty;
-                }
-
-                return ((AssemblyProductAttribute)attributes[0]).Product;
-            }
-        }
-
-        /// <summary>
-        /// Gets the AssemblyCopyright.
-        /// </summary>
-        public string AssemblyCopyright
-        {
-            get
-            {
-                object[] attributes = Assembly.GetExecutingAssembly().GetCustomAttributes(typeof(AssemblyCopyrightAttribute), false);
-                if (attributes.Length == 0)
-                {
-                    return string.Empty;
-                }
-
-                return ((AssemblyCopyrightAttribute)attributes[0]).Copyright;
-            }
-        }
-
-        /// <summary>
-        /// Gets the AssemblyCompany.
-        /// </summary>
-        public string AssemblyCompany
-        {
-            get
-            {
-                object[] attributes = Assembly.GetExecutingAssembly().GetCustomAttributes(typeof(AssemblyCompanyAttribute), false);
-                if (attributes.Length == 0)
-                {
-                    return string.Empty;
-                }
-
-                return ((AssemblyCompanyAttribute)attributes[0]).Company;
-            }
-        }
-        #endregion
-    }
-}
