--- ColumnSelector/Properties/AssemblyInfo.cs
@@ -1,36 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("ColumnSelector")]
-[assembly: AssemblyDescription("Reads input and returns a particular column.\n Table Select Icon courtesy of FatCow (http://www.fatcow.com/free-icons Farm [Fresh Icons - Table select column icon])")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Ace Olszowka")]
-[assembly: AssemblyProduct("ColumnSelector")]
-[assembly: AssemblyCopyright("Copyright © Ace Olszowka 2014-2022")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-// Setting ComVisible to false makes the types in this assembly not visible 
-// to COM components.  If you need to access a type in this assembly from 
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("e07fb290-e689-4c19-9640-e92d14c80dce")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version 
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers 
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.6")]
-[assembly: AssemblyFileVersion("1.0.0.6")]
