+++ Utilities.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.1.32210.238
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ColumnSelector", "ColumnSelector\ColumnSelector.csproj", "{87773F03-1E60-442D-93EC-80E9BB64839E}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ColumnSelector", "ColumnSelector\ColumnSelector.Windows\ColumnSelector.csproj", "{87773F03-1E60-442D-93EC-80E9BB64839E}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DictionaryFilter", "DictionaryFilter\DictionaryFilter.csproj", "{B565CD64-5C2E-47BA-9242-33BD23460F1F}"
 EndProject
